fix(operator): ensure correct bitwise operations in data() method
Add explicit casts to uint8_t in the data() method of the Operator class.
This prevents potential issues with implicit type conversions during
bitwise operations, ensuring the correct representation of operator data.

The change improves type safety and reduces the risk of unexpected
behavior in bitwise operations, particularly on systems where char
might be signed by default.

+++ src/operator.h
@@ -63,8 +63,9 @@ class Operator {
   }
 
   [[nodiscard]] constexpr uint8_t data() const noexcept {
-    return (data_.type << TYPE_BIT_POSITION) |
-           (data_.spin << SPIN_BIT_POSITION) | data_.orbital;
+    return static_cast<uint8_t>(data_.type << TYPE_BIT_POSITION) |
+           static_cast<uint8_t>(data_.spin << SPIN_BIT_POSITION) |
+           static_cast<uint8_t>(data_.orbital);
   }
 
  private:
